requirements.txt
supabase>=2.0.0
openai>=1.0.0
pdfplumber>=0.10.0
streamlit>=1.30.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
requests>=2.31.0
